INT to ONOS fail safe added
There was a scenario were a packet with INT headers coming to ingress from another switch, could end up selected to be cloned to the controller (CPU), and if the controller can't understand INT header that could cause the packet to be detected with corrupted data and be dropped.
To avoid this, a if condition was added at Egress to remove all the INT headers from the packet before it leaves from the switch to the controller.
The chance of this bug occuring is very low, but this prevents it from occuring.