Telemetry being sent to collector, mn image published, grafana and influxDB instructions added
Created instructions to prepare grafana and influxDB.
The custom stratum-mn image was renamed and published on docker hub as davidcc73/stratum_bmv2_x11_scapy_pip3

Switch's hosts ports are now hard coded in the topology.

Telemetry is being generated and removed from packet before going to final host, but the reports created after still have problems with the declared payload length.
Usage of SRv6 and INT at the same time is not completed